% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scatter.R
\docType{data}
\name{scatter}
\alias{scatter}
\title{Scatter data}
\format{
A data.frame including 100 observations and 2 variables
\describe{
\item{type}{character}
\item{cls}{integer represented ordered class}
\item{V1,V2}{numeric}
}
}
\usage{
scatter
}
\description{
This is data.
}
\keyword{datasets}
