% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostOLR.R
\docType{class}
\name{XGBMetricForOrderedLogit}
\alias{XGBMetricForOrderedLogit}
\title{Custom Objective/Evaluate Functions Class for XGBoost}
\description{
Custom Objective/Evaluate Functions Class for XGBoost

Custom Objective/Evaluate Functions Class for XGBoost
}
\details{
R6 Class defining custom Objective/Evaluate for XGBoost
}
\note{
Add -Inf/Inf both side of input x

x is sorted ascending
\code{criterion[1] = -Inf},
\code{criterion[2] = x[1]},
...
\code{criterion[nClass] = x[nClass - 1]},
\code{criterion[nClass + 1] = Inf}

x = 1 -> hit ratio is calculated by {|act - est| < x} / length(act)

grad, hess were referred to URL below.
https://www.slideshare.net/TakujiTahara/201200229-lt-dsb2019-ordered-logit-model-for-qwk-tawara#23

self$notchdiff_eval is set by set_notchdiff_eval
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Criterion}}{values between each labels for Ordered Logit Model}

\item{\code{notchdiff_eval}}{width of notch when evaluating hit ratio}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Criterion}}{values between each labels for Ordered Logit Model}

\item{\code{notchdiff_eval}}{width of notch when evaluating hit ratio}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set_criterion}{\code{XGBMetricForOrderedLogit$set_criterion()}}
\item \href{#method-return_criterion}{\code{XGBMetricForOrderedLogit$return_criterion()}}
\item \href{#method-set_notchdiff_eval}{\code{XGBMetricForOrderedLogit$set_notchdiff_eval()}}
\item \href{#method-return_notchdiff_eval}{\code{XGBMetricForOrderedLogit$return_notchdiff_eval()}}
\item \href{#method-pred_class_criterion}{\code{XGBMetricForOrderedLogit$pred_class_criterion()}}
\item \href{#method-pred_class_maxprob}{\code{XGBMetricForOrderedLogit$pred_class_maxprob()}}
\item \href{#method-obj_ordered_logit}{\code{XGBMetricForOrderedLogit$obj_ordered_logit()}}
\item \href{#method-eval_logloss}{\code{XGBMetricForOrderedLogit$eval_logloss()}}
\item \href{#method-eval_hitratio}{\code{XGBMetricForOrderedLogit$eval_hitratio()}}
\item \href{#method-clone}{\code{XGBMetricForOrderedLogit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_criterion"></a>}}
\if{latex}{\out{\hypertarget{method-set_criterion}{}}}
\subsection{Method \code{set_criterion()}}{
Set Criterions between each labels for Ordered Logit Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XGBMetricForOrderedLogit$set_criterion(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A double vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_criterion"></a>}}
\if{latex}{\out{\hypertarget{method-return_criterion}{}}}
\subsection{Method \code{return_criterion()}}{
Return criterions between each labels for Ordered Logit Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XGBMetricForOrderedLogit$return_criterion()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Criterion set in this class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_notchdiff_eval"></a>}}
\if{latex}{\out{\hypertarget{method-set_notchdiff_eval}{}}}
\subsection{Method \code{set_notchdiff_eval()}}{
Set notch width when evaluating hit ratio by eval_hitratio()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XGBMetricForOrderedLogit$set_notchdiff_eval(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{An integer, notch width when evaluate hit ratio}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_notchdiff_eval"></a>}}
\if{latex}{\out{\hypertarget{method-return_notchdiff_eval}{}}}
\subsection{Method \code{return_notchdiff_eval()}}{
Return notch width
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XGBMetricForOrderedLogit$return_notchdiff_eval()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
private$notchdiff_eval
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pred_class_criterion"></a>}}
\if{latex}{\out{\hypertarget{method-pred_class_criterion}{}}}
\subsection{Method \code{pred_class_criterion()}}{
Function to predict labels comparing preds and criterions (not maximum probability)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XGBMetricForOrderedLogit$pred_class_criterion(preds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{preds}}{A double vector, margin score from xgboost before logistic transformation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Integer label, if \code{[-Inf, criterion[1])} label is 1, else if \code{[criterion[1], criterion[2])}, label is 2, ...
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pred_class_maxprob"></a>}}
\if{latex}{\out{\hypertarget{method-pred_class_maxprob}{}}}
\subsection{Method \code{pred_class_maxprob()}}{
Function to predict labels at maximum probability
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XGBMetricForOrderedLogit$pred_class_maxprob(preds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{preds}}{A double vector, margin score from xgboost before logistic transformation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Integer label at max probability
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-obj_ordered_logit"></a>}}
\if{latex}{\out{\hypertarget{method-obj_ordered_logit}{}}}
\subsection{Method \code{obj_ordered_logit()}}{
Custom Objective Function of Ordered Logit Model
https://github.com/dmlc/xgboost/blob/master/R-package/demo/custom_objective.R
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XGBMetricForOrderedLogit$obj_ordered_logit(preds, dtrain)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{preds}}{A double vector, margin score at t-1}

\item{\code{dtrain}}{A xgb.DMatrix created by xgboost::xgb.DMatrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list(both double vector).
$grad gradient of loss function.
$hess hessian of loss function.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_logloss"></a>}}
\if{latex}{\out{\hypertarget{method-eval_logloss}{}}}
\subsection{Method \code{eval_logloss()}}{
Custom Evaluate Function of Ordered Logit Model (logloss)
https://github.com/dmlc/xgboost/blob/master/R-package/demo/custom_objective.R
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XGBMetricForOrderedLogit$eval_logloss(preds, dtrain)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{preds}}{A double vector, margin score}

\item{\code{dtrain}}{A xgb.DMatrix created by xgboost::xgb.DMatrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list(both double vector).
$metric name of metric.
$value value of metric.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_hitratio"></a>}}
\if{latex}{\out{\hypertarget{method-eval_hitratio}{}}}
\subsection{Method \code{eval_hitratio()}}{
Custom Evaluate Function of Ordered Logit Model (hit ratio)
https://github.com/dmlc/xgboost/blob/master/R-package/demo/custom_objective.R
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XGBMetricForOrderedLogit$eval_hitratio(preds, dtrain)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{preds}}{A double vector, margin score}

\item{\code{dtrain}}{A xgb.DMatrix created by xgboost::xgb.DMatrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list(both double vector).
$metric name of metric.
$value value of metric.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XGBMetricForOrderedLogit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
